Move swaps constants to the shared constants directory
