+++ shared/constants/swaps.js
