+++ ui/app/hooks/useCurrentAsset.js
@@ -2,7 +2,7 @@ import { useSelector } from 'react-redux';
 import { useRouteMatch } from 'react-router-dom';
 import { getTokens } from '../ducks/metamask/metamask';
 import { ASSET_ROUTE } from '../helpers/constants/routes';
-import { ETH_SWAPS_TOKEN_OBJECT } from '../helpers/constants/swaps';
+import { ETH_SWAPS_TOKEN_OBJECT } from '../../../shared/constants/swaps';
 
 /**
  * Returns a token object for the asset that is currently being viewed.
